pyproject.toml
[project]
name = "domainreviewer"
version = "0.1.0"
description = "A tool to review domains against OSINT sources like VirusTotal"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "Pillow>=10.0.0",
]
